Advance multiroot workspace support
a) This adds an ability to have a workspace-folder-specific checkpatch config.
This useful in case when you have projects with different requirements in your
multiroot workspace.

b) This adds a checkpatch-specific output channel (output pane of vscode)

c) This adds 'Clean Current Diagnostic' command to clean up problems pane manually